+++ streamingclient.go
@@ -62,7 +62,7 @@ func (client *StreamingClient) Authenticate(config *tls.Config, applicationKey, 
 	}
 
 	am := AuthenticationMessage{
-		ID:      int(rand.Uint64()),
+		ID:      int(rand.Uint32()),
 		Op:      "authentication",
 		AppKey:  applicationKey,
 		Session: sessionToken,
